Small update on Users Api & added /auth/facebook API
Updated a little bit the policies and checks upon user validation. On the future need to add the SAError Class to parse DB-related errors.
Created the Facebook Application and integrated the API for facebook logging in (accessed at /auth/facebook). If successful, the api creates a new user or connects (by email) an existing one to the facebook Profile. Also gets the profile picture and some other user-related info.